characters_service/requirements.txt
streamlit
flask
pdfplumber
requests
langchain
openai
python-dotenv
langchain_community